app/requirements.txt
gradio==4.7.1
numpy==1.23.4
tritonclient[grpc, http]==2.39.0
python-decouple==3.8